requirements.txt
keyboard
sounddevice
soundfile
google-genai
pyautogui
pyperclip
python-dotenv
pystray
Pillow
pyinstaller
faster-whisper
